// src/generated/types/docker-compose.ts
export type Include =
  | string
  | {
      /**
       * Path to the Compose application or sub-project files to include.
       */
      path?: string | ListOfStrings
      /**
       * Path to the environment files to use to define default values when interpolating variables in the Compose files being parsed.
       */
      env_file?: string | ListOfStrings
      /**
       * Path to resolve relative paths set in the Compose file
       */
      project_directory?: string
    }
/**
 * A list of string values.
 *
 * Items: A string value in the list.
 */
export type ListOfStrings = string[]
/**
 * Configuration for a service.
 *
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export type Service = ContainerSpec &
  WorkloadSpec & {
    deploy?: Deployment
    develop?: Development
    profiles?: ListOfStrings9
    /**
     * Restart policy for the service container. Options include: 'no', 'always', 'on-failure', and 'unless-stopped'.
     */
    restart?: string
    /**
     * Number of containers to deploy for this service.
     */
    scale?: number | string
    attach?: boolean | string
    /**
     * Specify a custom container name, rather than a generated default name.
     */
    container_name?: string
    /**
     * Specify a service which will not be manage by Compose directly, and delegate its management to an external provider.
     */
    provider?: {
      /**
       * External component used by Compose to manage setup and teardown lifecycle of the service.
       */
      type: string
      /**
       * Provider-specific options.
       */
      options?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string | number | boolean | (string | number | boolean)[]
      }
      /**
       * This interface was referenced by `undefined`'s JSON-Schema definition
       * via the `patternProperty` "^x-".
       */
      [k: string]: unknown
    }
    /**
     * Extend another service, in the current file or another file.
     */
    extends?:
      | string
      | {
          /**
           * The name of the service to extend.
           */
          service: string
          /**
           * The file path where the service to extend is defined.
           */
          file?: string
        }
    /**
     * Link to containers in another service. Either specify both the service name and a link alias (SERVICE:ALIAS), or just the service name.
     */
    links?: string[]
    /**
     * Link to services started outside this Compose application. Specify services as <service_name>:<alias>.
     */
    external_links?: string[]
    /**
     * Init containers to run to completion before the service container is started. Each step runs in its own ephemeral container, in declared order; a non-zero exit fails the bring-up of the service and its dependents.
     */
    pre_start?: PreStartHook[]
    /**
     * Commands to run after the container starts. If any command fails, the container stops.
     */
    post_start?: ServiceHook1[]
    /**
     * Commands to run before the container stops. If any command fails, the container stop is aborted.
     */
    pre_stop?: ServiceHook1[]
    [k: string]: unknown
  }
// ...
export type ListOfStrings9 = string[]
// ...
export type Network = {
  /**
   * Custom name for this network.
   */
  name?: string
  /**
   * Specify which driver should be used for this network. Default is 'bridge'.
   */
  driver?: string
  /**
   * Specify driver-specific options defined as key/value pairs.
   */
  driver_opts?: {
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^.+$".
     */
    [k: string]: string | number
  }
  /**
   * Custom IP Address Management configuration for this network.
   */
  ipam?: {
    /**
     * Custom IPAM driver, instead of the default.
     */
    driver?: string
    /**
     * List of IPAM configuration blocks.
     */
    config?: {
      /**
       * Subnet in CIDR format that represents a network segment.
       */
      subnet?: string
      /**
       * Range of IPs from which to allocate container IPs.
       */
      ip_range?: string
      /**
       * IPv4 or IPv6 gateway for the subnet.
       */
      gateway?: string
      /**
       * Auxiliary IPv4 or IPv6 addresses used by Network driver.
       */
      aux_addresses?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string
      }
      /**
       * This interface was referenced by `undefined`'s JSON-Schema definition
       * via the `patternProperty` "^x-".
       */
      [k: string]: unknown
    }[]
    /**
     * Driver-specific options for the IPAM driver.
     */
    options?: {
      /**
       * This interface was referenced by `undefined`'s JSON-Schema definition
       * via the `patternProperty` "^.+$".
       */
      [k: string]: string
    }
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown
  }
  /**
   * Specifies that this network already exists and was created outside of Compose.
   */
  external?:
    | boolean
    | string
    | {
        /**
         * @deprecated
         * Specifies the name of the external network. Deprecated: use the 'name' property instead.
         */
        name?: string
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^x-".
         */
        [k: string]: unknown
      }
  /**
   * Create an externally isolated network.
   */
  internal?: boolean | string
  /**
   * Enable IPv4 networking.
   */
  enable_ipv4?: boolean | string
  /**
   * Enable IPv6 networking.
   */
  enable_ipv6?: boolean | string
  /**
   * If true, standalone containers can attach to this network.
   */
  attachable?: boolean | string
  /**
   * Either a dictionary mapping keys to values, or a list of strings.
   */
  labels?:
    | {
        /**
         * Value for the key, which can be a string, number, boolean, or null.
         *
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` ".+".
         */
        [k: string]: string | number | boolean | null
      }
    | string[]
  /**
   * This interface was referenced by `undefined`'s JSON-Schema definition
   * via the `patternProperty` "^x-".
   */
  [k: string]: unknown
} | null
/**
 * Volume configuration for the Compose application.
 *
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export type Volume = {
  /**
   * Custom name for this volume.
   */
  name?: string
  /**
   * Specify which volume driver should be used for this volume.
   */
  driver?: string
  /**
   * Specify driver-specific options.
   */
  driver_opts?: {
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^.+$".
     */
    [k: string]: string | number
  }
  /**
   * Specifies that this volume already exists and was created outside of Compose.
   */
  external?:
    | boolean
    | string
    | {
        /**
         * @deprecated
         * Specifies the name of the external volume. Deprecated: use the 'name' property instead.
         */
        name?: string
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^x-".
         */
        [k: string]: unknown
      }
  /**
   * Either a dictionary mapping keys to values, or a list of strings.
   */
  labels?:
    | {
        /**
         * Value for the key, which can be a string, number, boolean, or null.
         *
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` ".+".
         */
        [k: string]: string | number | boolean | null
      }
    | string[]
  /**
   * This interface was referenced by `undefined`'s JSON-Schema definition
   * via the `patternProperty` "^x-".
   */
  [k: string]: unknown
} | null
/**
 * Configuration for a job. Jobs are containers that run to completion.
 *
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export type Job = ContainerSpec &
  WorkloadSpec & {
    profiles?: ListOfStrings10
    /**
     * Trigger conditions for the job. At least one trigger attribute must be declared. Setting manual to false forbids manual execution by an explicit run command.
     */
    triggers: {
      [k: string]: unknown
    } & {
      /**
       * Whether the job can be triggered manually by an explicit run command. Defaults to true; an explicit false forbids manual execution.
       */
      manual?: boolean | string
      /**
       * List of schedules for the job.
       */
      schedule?: (string | Schedule)[]
      /**
       * This interface was referenced by `undefined`'s JSON-Schema definition
       * via the `patternProperty` "^x-".
       */
      [k: string]: unknown
    }
    [k: string]: unknown
  }
/**
 * A list of unique string values.
 *
 * Items: A string value in the list.
 */
export type ListOfStrings10 = string[]

/**
 * The Compose file is a YAML file defining a multi-containers based application.
 */
export interface ComposeSpecification {
  /**
   * @deprecated
   * declared for backward compatibility, ignored. Please remove it.
   */
  version?: string
  /**
   * define the Compose project name, until user defines one explicitly.
   */
  name?: string
  /**
   * compose sub-projects to be included.
   */
  include?: Include[]
  /**
   * The services that will be used by your application.
   */
  services?: {
    [k: string]: Service
  }
  /**
   * Language models that will be used by your application.
   */
  models?: {
    [k: string]: Model
  }
  /**
   * Networks that are shared among multiple services.
   */
  networks?: {
    [k: string]: Network
  }
  /**
   * Named volumes that are shared among multiple services.
   */
  volumes?: {
    [k: string]: Volume
  }
  /**
   * Secrets that are shared among multiple services.
   */
  secrets?: {
    [k: string]: Secret
  }
  /**
   * Configurations that are shared among multiple services.
   */
  configs?: {
    [k: string]: Config
  }
  /**
   * Jobs are containers that run to completion.
   */
  jobs?: {
    [k: string]: Job
  }
  /**
   * This interface was referenced by `ComposeSpecification`'s JSON-Schema definition
   * via the `patternProperty` "^x-".
   */
  [k: string]: unknown
}
// ...
export interface Model {
  /**
   * Custom name for this model.
   */
  name?: string
  /**
   * Language Model to run.
   */
  model: string
  context_size?: number
  /**
   * Raw runtime flags to pass to the inference engine.
   */
  runtime_flags?: string[]
  /**
   * This interface was referenced by `Model`'s JSON-Schema definition
   * via the `patternProperty` "^x-".
   */
  [k: string]: unknown
}
/**
 * Secret configuration for the Compose application.
 *
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export interface Secret {
  /**
   * Custom name for this secret.
   */
  name?: string
  /**
   * Name of an environment variable from which to get the secret value.
   */
  environment?: string
  /**
   * Path to a file containing the secret value.
   */
  file?: string
  /**
   * Specifies that this secret already exists and was created outside of Compose.
   */
  external?:
    | boolean
    | string
    | {
        /**
         * Specifies the name of the external secret.
         */
        name?: string
        [k: string]: unknown
      }
  /**
   * Either a dictionary mapping keys to values, or a list of strings.
   */
  labels?:
    | {
        /**
         * Value for the key, which can be a string, number, boolean, or null.
         *
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` ".+".
         */
        [k: string]: string | number | boolean | null
      }
    | string[]
  /**
   * Specify which secret driver should be used for this secret.
   */
  driver?: string
  /**
   * Specify driver-specific options.
   */
  driver_opts?: {
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^.+$".
     */
    [k: string]: string | number
  }
  /**
   * Driver to use for templating the secret's value.
   */
  template_driver?: string
  /**
   * This interface was referenced by `Secret`'s JSON-Schema definition
   * via the `patternProperty` "^x-".
   */
  [k: string]: unknown
}